add streaming
